% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSingleTip.R
\name{addSingleTip}
\alias{addSingleTip}
\title{Add Tips to a Phylogenetic Tree}
\usage{
addSingleTip(phy, tip, tax, insert = "crown", ignore.monophyly,
  stem.edge = 0.5, quiet = FALSE)
}
\arguments{
\item{phy}{An object of class \code{\link{phylo}}.}

\item{tip}{A character string giving the name of the species to be added.}

\item{tax}{A data frame containing the taxonomic classification.}

\item{insert}{A character string indicating the positions where the 
species is to be inserted: \code{"crown"}, \code{"stem"}, 
\code{"randomly"}, or any unambigous abbreviation of these. 
This option will only have an effect if \code{phy} contains more than 
one congeneric of \code{tip}.}

\item{ignore.monophyly}{Logical, indicating if monophyly should be considered
when selecting the anchor point for \code{tips}.}

\item{stem.edge}{A real number greater than 0 and smaller than 1, which 
gives the fraction of the terminal branch length that will be assigned to 
the branch subtending the newly created MRCA of \code{tip} and its single 
congeneric species in \code{phy}. Will have no effect if \code{phy} 
contains more than one congeneric of \code{tip}.}

\item{quiet}{Logical, indicating if screen output should be 
suppressed.}
}
\value{
An object of class \code{\link{phylo}}.
}
\description{
Add tips (species) to a phylogenetic tree according 
to their taxonomic classification.
}
\seealso{
\code{\link{dbReadTaxonomy}} for reading a taxonomy table 
from the postgreSQL database.
}
