% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkBlocks.R
\name{checkBlocks}
\alias{checkBlocks}
\title{Summarize and Plot Aligment Blocks}
\usage{
checkBlocks(x, plot = TRUE, col, subset)
}
\arguments{
\item{x}{An object of class \code{\link{megapteraProj}}.}

\item{plot}{Logical, for suppressing the barplot.}

\item{col}{A vector of mode \code{"character"} giving two or more colors two
distinguish between alignment blocks in the barplot.}

\item{subset}{A vector of mode \code{"character"}, that can be used to choose
a subset of the total taxa available.}
}
\description{
Summarize and visualize alignment blocks with a barplot.
}
\seealso{
\code{\link{checkSpecLocus}}
}
