% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkIdentityCoverage.R
\name{checkIdentityCoverage}
\alias{checkIdentityCoverage}
\title{Histograms of Identity/Coverage per Locus}
\usage{
checkIdentityCoverage(megapteraProj)
}
\arguments{
\item{megapteraProj}{An object of class \code{\link{megapteraProj}}.}
}
\value{
None, as a side effect a PDF file is written to the working 
directory.
}
\description{
Print histograms of the distribution of identity and 
coverage for each locus in the database to a PDF file.
}
\seealso{
\code{\link{checkSpecLocus}},\code{\link{checkMissingSpec}}, 
\code{\link{checkSpecies}}.
}
