% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMissingSpec.R
\name{checkMissingSpec}
\alias{checkMissingSpec}
\title{Check for Missing Species}
\usage{
checkMissingSpec(x, provenance = ".", tag = "ingroup")
}
\arguments{
\item{x}{An object of class \code{\linkS4class{dbPars}} or
\code{\linkS4class{megapteraProj}}.}

\item{provenance}{A character string as used in the column \code{provenance}
the \code{acc_<locus>} tables in the postgreSQL database. Depending on your
actual pipeline setup these may be \code{"ncbi"}, \code{"ncbi annotated"},
or \code{"bold"} as well as any user-defined strings.}

\item{tag}{A character string as used in the column \code{tag} in the
\code{taxonomy} table in the postgreSQL database. Depending on your choices
for parameter values in \code{\link{taxon}} or \code{\link{taxonGuidetree}}
these may be \code{"ingroup (NCBI)"}, \code{"extended ingroup (NCBI)"},
\code{"outgroup (NCBI)"}, or \code{"extended outgroup (NCBI)"} as well as
any user-defined strings.}
}
\value{
A data frame with two columns: 
\describe{ 
  \item{spec}{Names of missing species} 
  \item{status}{Why species is missing; might be (1)
  \code{not.on.genbank} (species for which no sequences were found on
  GenBank), (2) \code{not.selected} (species that were not selected due to
  below-threshold identity or coverage), or (3) \code{not.aligned} (species
  that were lost during the alignment process; happens rarely)}
  }
}
\description{
Create a table of species that failed to pass steps in the
  pipeline.
}
\seealso{
\code{\link{checkSpecLocus}} to create a barplot of the species number per locus.
}
