% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSpecies.R
\name{checkSpecies}
\alias{checkSpecies}
\title{Get Information about Species}
\usage{
checkSpecies(megProj, spec)
}
\arguments{
\item{megProj}{An object of class \code{\link{megapteraProj}}.}

\item{spec}{A character string giving the name of a species.}
}
\value{
A data frame, as a side effect a summary message is printed on the screen.
}
\description{
Returns information about the 'fate' of a single species along the pipeline.
}
\seealso{
\code{\link{checkMissingSpec}} to create a list of species that were missed/lost during consecutive steps of the pipeline.
}
