% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSubset.R
\name{checkSubset}
\alias{checkSubset}
\title{Check A Taxonomic Subset}
\usage{
checkSubset(x, subset)
}
\arguments{
\item{x}{Either an object of class \code{\link{megapteraProj}} or a data
frame as returned by \code{\link{dbReadTaxonomy}}.}

\item{subset}{A list of species names. The list's elements can list more than
one species; in this case the first name is taken as accepted name and the
others as synonms.}
}
\value{
A data frame with two columns: 
\describe{ 
\item{\code{taxon}}{the
  accepted name of the taxon} 
\item{\code{present}}{the status of the taxon,
  i.e. present or not}}
}
\description{
Checks the availability of a set of species in a megaptera
  project.
}
\seealso{
\code{\link{checkSpecies}}
}
