% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseSingles.R
\name{collapseSingles}
\alias{collapseSingles}
\title{Collapse Singleton Edges}
\usage{
collapseSingles(phy)
}
\arguments{
\item{phy}{An object of class \code{"\link{phylo}"}.}
}
\value{
An object of class \code{"\link{phylo}"}.
}
\description{
Collapse singleton edges in a phylogenetic tree. This function
  is an alternative for collapse.singleton from the \strong{ape} package that
  will also work when \code{ape::reorder(phy)} throws an error. (And this
  tends to happen often when taxonomic relationships are converted to trees.)
}
\seealso{
\code{\link{taxdump2phylo}}, \code{\link{comprehensiveGuidetree}}
}
