% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehensiveGuidetree.R
\name{comprehensiveGuidetree}
\alias{comprehensiveGuidetree}
\title{Comprehensive Guide Tree}
\usage{
comprehensiveGuidetree(megProj, tip.rank, subset)
}
\arguments{
\item{megProj}{An object of class \code{\link{megapteraProj}}.}

\item{tip.rank}{A character string giving the rank of the tips (e.g. 
\code{spec}, \code{gen}, ...)}

\item{subset}{A subset of species names the taxonomy should be limited to. 
This can be a DNA alignment of class \code{DNAbin}, a phylogenetic tree of class
\code{phylo}, a data frame or a character vector listing the species names.}
}
\value{
An object of class \code{\link{phylo}}.
}
\description{
Creates a complete (or comprehensive) guide tree for the
  alignment of sequences based on the underlying taxonomic
  classification. If available (see \code{\link{taxonGuidetree}}), a
  user-defined guide tree of higher ranks will be incorporated to enhance the
  resolution of the comprehensive guide tree.
}
\details{
It can be desirable to include a user-defined guide tree for two 
  reasons. First, one might favor another phylogenetic hypothesis than 
  inherent in the NCBI classification. Second, no classification can provide 
  intra-rank phylogenetic relationships, but an explicit phylogeny does. 
  Hence, via the user-defined guide tree, it is possible to add more 
  resolution to the guide tree topology.
}
\seealso{
\code{\link{dbReadTaxonomy}} for reading a taxonomic classification 
  from the postgreSQL database.
}
