% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbMaxGIPerSpec.R
\name{dbMaxGIPerSpec}
\alias{dbMaxGIPerSpec}
\title{Limit Numbers of Sequences per Species}
\usage{
dbMaxGIPerSpec(megProj, max.gi.per.spec = 10, prefer = "longest", taxon)
}
\arguments{
\item{megProj}{An object of class \code{\link{megapteraProj}}.}

\item{max.gi.per.spec}{An integer giving the maximum number of sequences per
species to be used by the pipeline.}

\item{prefer}{A character string indicating under what criterion the
sequences will be choosen; available are \code{"longest"} (default),
\code{"shortest"}, \code{"most frequent length"} and \code{"random"}.}

\item{taxon}{A character string giving one or more taxon names for which the
number of sequences will be limited. If left missing, all taxon names found
in the database are handled.}
}
\description{
Some species (e.g. model organism) have thousands or more
  sequences of a single locus on GenBank. Much of this genetic information is
  redundant in a phylogenetic context, so \strong{megaptera} internally
  limits the number of sequences per species to 10. This number can be
  changed by the user via with \code{megapteraPars} or \code{dbMaxGIPerSpec},
  the latter function allowing more fine-tuning (see Details).
}
\details{
After calling \code{dbMaxGIPerSpec} the pipeline must be rerun from
  \code{\link{stepC}} onwards.

  \emph{To do: describe parameters!}
}
\seealso{
\code{\link{megapteraPars}}
}
