% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbPars-class.R
\docType{class}
\name{dbPars-class}
\alias{dbPars-class}
\title{Class "dbPars"}
\description{
S4 class for database parameters for a megaptera project, as
  defined with \code{\link{megapteraProj}}.
}
\section{Slots}{

\describe{
\item{\code{host}}{A vector of mode \code{"character"}, giving the name of the host,
most often this will be \code{"localhost"}.}

\item{\code{port}}{Numeric, giving the port number, most often \code{5432}.}

\item{\code{dbname}}{A vector of mode \code{"character"}, giving the name of the
database.}

\item{\code{user}}{A vector of mode \code{"character"}, giving the name of the user.}

\item{\code{password}}{A vector of mode \code{"character"}, giving the password.}
}}

\seealso{
\code{\link{dbPars}}, \code{\link{locus}}, and
  \code{\link{megapteraPars}} for defining of database parameters, loci, and
  the pipeline's parameters, respectively, and \code{\link{megapteraProj}}
  for the bundling of input data.
}
