% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbPars.R
\name{dbPars}
\alias{dbPars}
\title{Set Database Parameters}
\usage{
dbPars(host = "localhost", port = 5432, dbname, user = "postgres", password)
}
\arguments{
\item{host}{A vector of mode \code{"character"}, giving the name of the host,
most often this will be \code{"localhost"}.}

\item{port}{Numeric, giving the port number, most often \code{5432}.}

\item{dbname}{A vector of mode \code{"character"}, giving the name of the
database.}

\item{user}{A vector of mode \code{"character"}, giving the name of the user.}

\item{password}{A vector of mode \code{"character"}, giving the password.}
}
\value{
An object of class \code{\linkS4class{dbPars}}.
}
\description{
Sets the connection parameters for a PostgreSQL database and, as
  a side effect, creates the database, if it does not yet exist.
}
\details{
\bold{megaptera} stores all data internally in a PostgreSQL
  database. Therefore, \code{dbPars} represents the first step in setting up a megaptera
  project pipeline. See \code{\link[DBI]{dbConnect}} and
  \code{\link{PostgreSQL}} for further details about the connection
  parameters and procedure.
}
\references{
See the documentation at the PostgreSQL Web site
  \url{http://www.postgresql.org} for details.
}
\seealso{
\code{\linkS4class{dbPars}} for the class' description;
  \code{\link{taxon}}, \code{\link{locus}}, and \code{\link{megapteraPars}}
  for defining of taxa, loci, and the pipeline's parameters, respectively;
  and \code{\link{megapteraProj}} for the bundling of input data.
}
