% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbReadMSA.R
\name{dbReadMSA}
\alias{dbReadMSA}
\title{Read Sequences from Database}
\usage{
dbReadMSA(x, taxon, regex = TRUE, ignore.excluded = TRUE,
  blocks = "ignore")
}
\arguments{
\item{x}{An object of class \code{\link{megapteraProj}}.}

\item{taxon}{A vector of mode \code{"character"} used to choose a subset of
available taxa. This can be either one or more taxon names or a regular
expression.}

\item{regex}{Logical: if \code{TRUE}, the string given via \code{taxon} will
be interpreted as a regular expression (see \code{\link{regex}}).}

\item{ignore.excluded}{\emph{Currently unused}.}

\item{blocks}{\emph{Currently unused}.}
}
\value{
An object of class \code{\link{DNAbin}}.
}
\description{
Reads selected and assembled sequences (see \code{\link{stepF}}
  and subsequent steps) from the database, specifically from the table
  'species_sequences'.
}
\seealso{
\code{\link{dbReadDNA}} and \code{\link{dbWriteDNA}}
}
