% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbRenameSpecies.R
\name{dbRenameSpecies}
\alias{dbRenameSpecies}
\title{Rename a Species}
\usage{
dbRenameSpecies(x, from, to)
}
\arguments{
\item{x}{An object of class \code{\link{megapteraProj}}.}

\item{from}{A vector of mode \code{"character"} giving the species name that should be replaced.}

\item{to}{A vector of mode \code{"character"} giving the new species name.}
}
\value{
None, \code{dbRenameSpecies} is called for its side effect.
}
\description{
Rename a species in all tables of the database.
}
\seealso{
\code{\link{dbDeleteSpec}}, \code{\link{dbExcludeSpec}}, \code{\link{dbIncludeSpec}}
}
