% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSummaryTaxonomy.R
\name{dbSummaryTaxonomy}
\alias{dbSummaryTaxonomy}
\title{Numerical Summary of Taxonomy}
\usage{
dbSummaryTaxonomy(megProj)
}
\arguments{
\item{megProj}{An object of class
\code{\linkS4class{megapteraProj}}.}
}
\description{
Returns the number of total entries (nodes) in the taxonomy as
  well as the number of species, genera, families and orders.
}
\seealso{
\code{\link{ncbiTaxonomy}} for downloading the NCBI taxonomy;
  \code{\link{stepA}} for extracting the taxonomic information relevant for a
  given megaptera project; \code{\link{dbUpdateTaxonomy}} and
  \code{\link{dbReadTaxonomy}} for storing and retrieving taxonomic
  information in a megaptera project database.
}
