% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findInFlatfile.R
\name{findInFlatfile}
\alias{findInFlatfile}
\title{Find Taxa and Accession Numbers in GenBank Flatfiles}
\usage{
findInFlatfile(file, taxon = NULL, acc = NULL)
}
\arguments{
\item{file}{A character string giving the name of the GenBank flatfile.}

\item{taxon}{A character string giving the name of the taxon.}

\item{acc}{A character string giving the accession number.}
}
\value{
Logical, returns \code{TRUE}, if \code{taxon} is included with one or
  more entries in \code{file}.
}
\description{
Checks if any entries of a taxon are present in a given GenBank
  flatfile based on regular expressions. The motivation is to save computation
  time by not parsing files that do not include the taxon of interest.
}
\keyword{internal}
