% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findRoot.R
\name{findRoot}
\alias{findRoot}
\title{Lineage Down to the Root}
\usage{
findRoot(x, what)
}
\arguments{
\item{x}{An object of class \code{\link{megapteraProj}}.}

\item{what}{A character string, either \code{"ingroup"}, \code{"outgroup"}, or \code{"both"}.}
}
\value{
A data frame with three columns:
 \item{id}{the unique identifier of the taxon}
 \item{taxon}{the scientific name of the taxon}
 \item{rank}{the rank of the taxon}
}
\description{
Finds the lineage from one taxon, or the most recent common
  ancestor of several taxa, down to the root of the Tree of Life.
}
