% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbflat2db.R
\name{gbflat2db}
\alias{gbflat2db}
\title{Read GenBank Flatfiles}
\usage{
gbflat2db(file, taxa = NULL, megProj)
}
\arguments{
\item{file}{A vector of mode \code{"character"} the path to the GenBank flatfile.}

\item{taxa}{A vector of mode \code{"character"} giving the set of taxon names
that will be selected from the availbale taxon names (ORGANISM entry in
flatfile).}

\item{megProj}{An object of class \code{\link{megapteraProj}}.}
}
\description{
Read a GenBank flatfile, optionally filter its contents according to
  organism name and sequence length and return as an object of class
  \code{\link{DNAbin}.}
}
