% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbflat2DNAbin.R
\name{gbflat2DNAbin}
\alias{gbflat2DNAbin}
\title{Read GenBank Flatfiles}
\usage{
gbflat2DNAbin(x, taxa)
}
\arguments{
\item{x}{A vector of mode \code{"character"} containg the contents of the
GenBank flatfile.}

\item{taxa}{A vector of mode \code{"character"} giving the set of taxon names
that will be selected from the availbale taxon names (ORGANISM entry in
flatfile).}
}
\description{
Read a GenBank flatfile, filter it contents according to
  organism name and sequence length and return as an object of class
  \code{\link{DNAbin}.}
}
