% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get3LetterCode.R
\name{get3LetterCode}
\alias{get3LetterCode}
\title{GenBank Division Code}
\usage{
get3LetterCode(x)
}
\arguments{
\item{x}{An object of class \code{\link{megapteraProj}}.}
}
\value{
A vector of mode \code{"character"} containing three-letter
  abreviations of the GenBank Division Code.
}
\description{
Get the GenBank Division Code based on the current taxonomy.
}
\references{
\url{https://www.ncbi.nlm.nih.gov/genbank/htgs/divisions/}
}
