% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higherTaxaOnEdges.R
\name{higherTaxaOnEdges}
\alias{higherTaxaOnEdges}
\title{Higher Taxa on Edges}
\usage{
higherTaxaOnEdges(phy, tax)
}
\arguments{
\item{phy}{An object of class \code{\link{phylo}}.}

\item{tax}{A data frame containing a taxonomy table in parent-child format.}
}
\value{
A data frame with two columns: 
\describe{ 
  \item{taxon}{the name of the taxon} 
  \item{edge}{the number of the edge in \code{phy}} 
}
}
\description{
Create a vector of taxon names to label edges in a phylogenetic 
  tree with \code{\link{edgelabels}}.
}
\seealso{
\code{\link{dbReadTaxonomy}}
}
