\name{icContingency}
\alias{icContingency}
\title{
Cumulative Contingency for Identity and Coverage
}
\description{
Calculate a matrix that contains the number of species selected for inclusion in a dataset for different threshold combinations of identity and coverage.
}
\usage{
icContingency(x, step = 0.05)
}
\arguments{
  \item{x}{
an object of class \code{\link{megapteraProj}}.
}
  \item{step}{
a real number in the interval ]0, 1[.
}
}
%\details{}
\value{
a matrix that contains the number of species selected for inclusion in a dataset for different threshold combinations of identity and coverage.
}
%\references{}
\author{Christoph Heibl}
\seealso{
\code{\link{locus}}, \code{\link{locusRef}}
}
%\examples{}

