% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indet.strings.R
\name{indet.strings}
\alias{indet.strings}
\title{Identify Undetermined Species}
\usage{
indet.strings(exclude.hybrids = TRUE, collapse = FALSE, SQL = FALSE)
}
\arguments{
\item{exclude.hybrids}{Logical, indicating if hybrids should be also identified by
the set of regular expressions.}

\item{collapse}{Logical, if \code{TRUE} the regular expressions are collected
in one single characters string using the \code{|} operator.}

\item{SQL}{Logical, \emph{currently unused}.}
}
\description{
Provides a set of regular expressions (\code{\link{regex}}) that
  identify undetermined names as they will appear in the NCBI Taxonomy.
}
\keyword{internal}
