% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.ingroup.R
\name{is.ingroup}
\alias{is.ingroup}
\title{Does a Taxon Belong to Ingroup?}
\usage{
is.ingroup(megProj, taxon)
}
\arguments{
\item{megProj}{An object of class \code{\link{megapteraProj}}.}

\item{taxon}{A character vector containing taxon names.}
}
\value{
A vector of logical values.
}
\description{
Check if taxon belong to the ingroup as defined by \code{\link{taxon}}.
}
