\name{megaptera-internal}

% these function deserve proper documentation
\alias{giConsensus}
\alias{megaptera2Rmarkdown}
\alias{plotPhyloWithClassification}
\alias{stepBOLD}
\alias{supermatrix.mrbayes}


\alias{conn}
\alias{cropToReference} % compareToRef
\alias{tax}
\alias{loci}
\alias{pars}
\alias{addNbAcc}
\alias{addRanks} % ncbiTaxonomy
\alias{alignGenus} % stepG
\alias{alignSisterClades} % stepG
\alias{alignSpecies} % stepC
\alias{as.Linnean} % stepE
\alias{boxplotMSA}
\alias{checkBlocks}
\alias{checkExcludedSpec}
\alias{checkStatus}
\alias{check.TaxAgainstPhy}
\alias{compareToRef} % stepE
\alias{compare.phylo.backbone}
\alias{coverage}
\alias{dbAutoFillTaxonomy}
\alias{dbChangeSpecName}
\alias{dbconnect}
\alias{dbChooseRef}
\alias{dbExcludeIndet}
\alias{dbMaxGIPerSpec}
\alias{dbSummaryMSA}
\alias{dbTableNames}
\alias{dbUpdateSurrogate}
\alias{dbWriteMSA}
\alias{dbPReadDNA}
\alias{dbReadReference}
\alias{dbUpdateLocus}
\alias{dbUpdateReference} % stepD
\alias{dbWriteDNAbin}
\alias{dbWriteString}
\alias{downloadSequences}
\alias{EFetchHigherTaxon}
\alias{EFetchLocus}
\alias{EFetchXML}
\alias{esearch}
\alias{extractIGS}
\alias{extractLocus}
\alias{filter.alignment}
\alias{findRoot} % ncbiGenome
\alias{fixTaxonomy} % deserves explicit doc
\alias{force.mp}
\alias{fuzzy.match}
\alias{getGeneRange}
\alias{getLineage}
\alias{getMaxDist}
\alias{htmlTable} % megaptera2Rmarkdown
\alias{indet.strings} % stepB, dbUpdateTaxonomy, ncbiGenome
\alias{insert.tip.phylo}
\alias{is.alignable}
\alias{is.Linnean} % stepE
\alias{MAD}
\alias{md5} % stepF
\alias{mafft.merge}
\alias{markersGenbank}
\alias{maxDist}
\alias{maxDistMPI}
\alias{meanPairwiseIdentity}
\alias{megapteraHeatmap}
\alias{myDist}
\alias{ncbiLineage}
\alias{NCBI.wrap}
\alias{nucleotideOverlap}
\alias{removeIsolatedTails} % stepD
\alias{revCompTest}
\alias{robustXMLparse} % downloadSequences
\alias{select.seqs}
\alias{slog}
\alias{specCons} % stepD
\alias{speciesConsensus}
\alias{sortRanks}
\alias{splitAlignment} % stepH
\alias{splitGiTaxon}
\alias{sql.conform}
\alias{sqlTaxonomyHeader}
\alias{sql.wrap}
\alias{stepAA}
\alias{summary.sproj}
\alias{tab}
\alias{tax2tree}
\alias{taxonomy2html} % megaptera2Rmarkdown
\alias{term}
\alias{valid.NCBI.taxon}
\alias{whereToInsert}
\alias{wrapSQL}
\alias{write.dna.spectable}

\title{Internal MEGAPTERA Functions}
\description{Internal \pkg{megaptera} functions.}
\note{
These are internal functions in \code{megaptera} and are not intended to be called by the user.
}
\keyword{internal}
