% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/megaptera2Rmarkdown.R
\name{megaptera2Rmarkdown}
\alias{megaptera2Rmarkdown}
\title{RMarkdown Report}
\usage{
megaptera2Rmarkdown(x, file)
}
\arguments{
\item{x}{An object of class \code{\link{megapteraProj}}.}

\item{file}{A character vector giving a file name.}
}
\value{
None, \code{megaptera2Rmarkdown} is called for its side effect.
}
\description{
Creates a status report for a megaptera Project using RMarkdown.
}
\seealso{
\code{\link{checkSpecLoc}}, \code{\link{checkSpecies}}
}
