% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/megapteraInit.R
\name{megapteraInit}
\alias{megapteraInit}
\title{Initialize MEGAPTERA}
\usage{
megapteraInit(path)
}
\arguments{
\item{path}{A vector of mode \code{"character"} that gives the directory
where megaptera should store data and results.}
}
\value{
None, \code{megapteraInit} is called for its side effects.
}
\description{
Creates a consist system of folders where data and metadata are
  stored.
}
\details{
\code{"megapteraInit"} creates a file system that is designed to
  keep memory requirements and execution times at the necessary minimum
  during subsequent runs of the MEGAPTERA pipeline. Individula projects can
  be created with \code{\link{megapteraProj}}.
}
\seealso{
\code{\link{megapteraProj}}, \code{\link{ncbiTaxonomy}}
}
