% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfPhyloA0.R
\name{pdfPhyloA0}
\alias{pdfPhyloA0}
\title{Plot Large Phylogenies}
\usage{
pdfPhyloA0(phy, file = "bigtree.pdf", view = FALSE, save = TRUE)
}
\arguments{
\item{phy}{An object of class \code{\link{phylo}}.}

\item{file}{A vector of mode \code{"character"} giving a filename (and path)
for the PDF file.}

\item{view}{Logical, if \code{TRUE}, the PDF will be opened in the default
PDF viewer, but nothing is saved.}

\item{save}{Logical, if \code{TRUE}, the PDF is saved to \code{file}.}
}
\value{
None, \code{slicePhylo} is called for its side effect of generating a
  PDF file.
}
\description{
Create a PDF file of a large phylogeny.
}
