% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg2DNAbin.R
\name{pg2DNAbin}
\alias{pg2DNAbin}
\title{Convert SQL Query Result to DNAbin}
\usage{
pg2DNAbin(pg, label, confid.scores, col.confid = 0, row.confid = 0)
}
\arguments{
\item{pg}{A data frame as returned by \code{dbGetQuery}.}

\item{label}{A vector of mode \code{"character"} giving the attributes for
constructing the sequences labels, e.g. \code{"taxon"}, \code{"acc"}, or both.}

\item{confid.scores}{A vector of mode \code{"character"} specifying if and
how confidence scores will be returned. Use \code{"ignore"} to omit
confidence scores. Other options are \code{"all"}, \code{"row.means"}, and
\code{"col.means"}. Confidence scores are returned as an attribute and can
be accessed by \code{attr(obj, "cs")}.}

\item{col.confid}{A real number in the interval \code{[0, 1]} giving
the confidence threshold for alignment columns (i.e. nucletide positions); only rows
(i.e. taxa, sequences) or columns  scoring equal
or greater to \code{min.confid} will be selected.}

\item{row.confid}{A real number in the interval \code{[0, 1]} giving
the confidence threshold for alignment rows (i.e. taxa, sequences); only rows  scoring equal
or greater to \code{min.confid} will be selected.}
}
\description{
Convert a data frame as returned by \code{dbGetQuery} to an
  object of class \code{\link{DNAbin}}, possibly with confidence scores.
}
