% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectMSA.R
\name{selectMSA}
\alias{selectMSA}
\title{Selection of Multiple Sequence Alignments}
\usage{
selectMSA(
  megProj,
  min.n.seq = 3,
  blocks = "split",
  row.confid = 0,
  col.confid = 0,
  trim.ends = 0,
  locus.coverage = 0,
  global.coverage = 0,
  subset.locus = NULL,
  subset.species = NULL,
  exclude.locus = NULL,
  exclude.species = NULL,
  core.locus = NULL,
  core.species = NULL,
  best.sampled.congeneric = FALSE,
  equal.taxon.sets = TRUE,
  use.outgroup = TRUE,
  protect.outgroup = FALSE,
  squeeze.outgroup = NULL
)
}
\description{
Selection of Multiple Sequence Alignments
}
