\name{setLocus}
\alias{setLocus}
\title{Set locus in megapteraProj}
\description{
Sets the \code{locus} slot in \code{\link{megapteraProj}}.
}
\usage{
setLocus(x, locus)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
an object of class \code{megapteraProj}.
}
  \item{locus}{
an object of class \code{locus}.
}
}
%\details{}
\value{
an object of class \code{megapteraProj}.
}
%\references{}
\author{
Christoph Heibl
}
\seealso{
\code{\link{megapteraProj}}, \code{\link{locus}}.
}
%\examples{}

