% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slicePhylo.R
\name{slicePhylo}
\alias{slicePhylo}
\title{Plot Very Large Phylogenies}
\usage{
slicePhylo(phy, file)
}
\arguments{
\item{phy}{An object of class \code{\link{phylo}}.}

\item{file}{A vector of mode \code{"character"} giving a filename (and path)
for the PDF file.}
}
\value{
None, \code{slicePhylo} is called for its side effect of
  generating one or more PDF file.
}
\description{
Cut a phylogeny into suitable slices (at least try to - result
  will always depend on topoloy) and create a series of PDF files showing the
  individual slices. This technique is meant to facilitate the graphical
  inspection of very large phylogenies.
}
