% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speclist2taxdump.R
\name{speclist2taxdump}
\alias{speclist2taxdump}
\title{Utilities for NCBI Taxdump}
\usage{
speclist2taxdump(species)
}
\arguments{
\item{species}{A vector of mode \code{"character"} giving subset.}
}
\description{
Convert a species list into a taxonomy table in parent-child format.
}
\seealso{
\code{\link{dbReadTaxonomy}},
  \code{\link{taxdumpChildren}},\code{\link{taxdumpLineage}},
  \code{\link{taxdumpAddNode}}, \code{\link{taxdump2phylo}}.
}
