% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectable2taxdump.R
\name{spectable2taxdump}
\alias{spectable2taxdump}
\title{Utilities for NCBI Taxdump}
\usage{
spectable2taxdump(x, root = "root", root.rank = "no rank", ranks = "asis")
}
\arguments{
\item{x}{A vector of mode \code{"character"} giving a set of species or
genus names.}

\item{root}{A character string giving the taxon name of the root node.}

\item{root.rank}{A character string giving the rank of the root node.}

\item{ranks}{A vector of mode \code{"character"} giving the ranks (= column
names) that will be extracted from \code{x}. Ranks must be ordered from
high to low (e.g. \code{c("family", "genus", "species")}).}
}
\description{
Convert a list of species or genera into a taxonomy table in
  parent-child format.
}
\seealso{
\code{\link{dbReadTaxonomy}},
  \code{\link{taxdumpChildren}},\code{\link{taxdumpLineage}},
  \code{\link{taxdumpAddNode}}, \code{\link{taxdump2phylo}}.
}
