% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepB.R
\name{stepB}
\alias{stepB}
\title{Step B: Search and Download Sequences}
\usage{
stepB(x, update.seqs = "no")
}
\arguments{
\item{x}{An object of class \code{\link{megapteraProj}}.}

\item{update.seqs}{A character string determining the behaviour of 
\code{stepB} when it is run repeatedly on the same locus/taxon 
combination: \code{"no"} means that only sequences that have been made 
available on GenBank since the last execution \code{stepB} will be 
downloaded; \code{"all"} means the database table with all sequences 
will be removed and all sequences will be downloaded de novo. Naturely, 
this option has no effect if \code{stepB} is called for the first time.}
}
\value{
None. \code{stepB} is called for its side effects: (1) strings of 
DNA sequences with attribute data are stored in a pgSQL database; (2) a log 
file is written to the current working directory.
}
\description{
For any given project (see \code{\link{megapteraProj}}), 
this step searches the Nucleotide database on GenBank, downloads all 
sequences and stores them in a postgreSQL database table.
}
\details{
All accessions are stored under their species name as appearing 
in the \emph{organism} field at GenBank, but information about infrageneric 
ranks is stripped off the taxon names before they are stored in the database.
}
\references{
\url{http://www.ncbi.nlm.nih.gov/books/NBK25501/}
}
\seealso{
\code{\link{megapteraProj}}; \code{\link{stepA}} for the preceeding 
and \code{\link{stepC}} for the subsequent step; \code{\link{stepBX}} for the 
addition of external sequences to the database.
}
