% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepBLAST.R
\name{stepBLAST}
\alias{stepBLAST}
\title{stepBLASTN: Filter Homologous Sequences}
\usage{
stepBLAST(x)
}
\arguments{
\item{x}{An object of class \code{\link{megapteraProj}}.}
}
\description{
Filters homologous sequences using the BLASTB tool
}
\references{
NCBI BLAST: \url{https://blast.ncbi.nlm.nih.gov/Blast.cgi}
}
