\name{stepC}
\alias{stepC}
\title{
Step C: Align Conspecific Sequences
}
\description{
Align conspecific sequences. In this context, `conspecific' refers to sequences whose names as given by the \emph{organism} field spell absolutely identical.
}
\usage{
stepC(x)
}
\arguments{
  \item{x}{an object of class \code{\link{megapteraProj}}}
}
%\details{}
%\value{}
%\references{}
\author{
Christoph Heibl
}
\note{
The function is under development and the results should always be checked.
}
\seealso{
\code{\link{megapteraProj}}; \code{\link{stepB}} for the preceeding and \code{\link{stepD}} for the subsequent step.
}
%\examples{}
