\name{stepD}
\alias{stepD}
\title{
Step D: Calculate Benchmark Sequences
}
\description{
In order to evaluate sequence identity and quality, each sequence will be compared to a `benchmark sequence', which is constructed in \code{stepD}.
}
\usage{
stepD(x)
}
\arguments{
    \item{x}{an object of class \code{\link{megapteraProj}}.
    }
}
%\details{}
%\value{}
%\references{}
\author{
Christoph Heibl
}
\note{
The function is under development and the results should always be checked.
}
\seealso{
\code{\link{megapteraProj}}; \code{\link{stepC}} for the preceeding and \code{\link{stepE}} for the subsequent step.
}
%\examples{}