% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepGBLOCKS.R
\name{stepGBLOCKS}
\alias{stepGBLOCKS}
\title{Detect Homology Uncertainty in Alingment}
\usage{
stepGBLOCKS(x)
}
\arguments{
\item{x}{An object of class \code{\link{megapteraProj}}.}
}
\value{
None, \code{stepGBLOCKS} is called for its side effects.
}
\description{
Use GBLOCKS to detect positions of uncertain homology and remove
  them from the alignment.
}
\seealso{
\code{\link{megapteraProj}}; \code{\link{stepMAFFT}} for the
  preceeding step and \code{\link{stepH}} for the subsequent step.
}
