% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepGenBankFTP.R
\name{stepGenBankFTP}
\alias{stepGenBankFTP}
\title{Extract Records from GenBank Flatfiles}
\usage{
stepGenBankFTP(x, update)
}
\arguments{
\item{x}{An object of class \code{\link{megapteraProj}}}

\item{update}{A vector of mode \code{"character"} giving the name of a taxon.}
}
\description{
Extract sequences of study taxa from GenBank flatfile.
}
