% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepMAFFT.R
\name{stepMAFFT}
\alias{stepMAFFT}
\title{STEP G: MAFFT Alignment}
\usage{
stepMAFFT(
  x,
  method = "auto",
  maxiterate = 0,
  op = 1.53,
  ep = 0,
  use.guidetree = TRUE
)
}
\arguments{
\item{x}{An object of class \code{\link{megapteraProj}}.}

\item{method}{A character string giving the alignment method. Available
accuracy-oriented methods for less than 200 sequences are
\code{"localpair"}, \code{"globalpair"}, and \code{"genafpair"};
\code{"retree 1"} and \code{"retree 2"} are for speed-oriented alignment.
The default is \code{"auto"}, which lets MAFFT choose an appropriate
alignment method.}

\item{maxiterate}{An integer giving the number of cycles of iterative
refinement to perform. Possible choices are \code{0}: progressive method,
no iterative refinement (default); \code{2}: two cycles of iterative
refinement; \code{1000}: at most 1000 cycles of iterative refinement.}

\item{op}{A numeric giving the \code{gap opening penalty} at group-to-group
alignment; default 1.53.}

\item{ep}{A numeric giving the offset value, which works like \code{gap
extension penalty}, for group-to-group alignment; default 0.0, but 0.123 is
recommended if no long indels are expected.}

\item{use.guidetree}{Logical, indicating whether to use megaptera's guide
tree or not. In the latter case, MAFFT's internal guide tree is used.}
}
\value{
None, \code{stepMAFFT} is called for its side effect.
}
\description{
Use MAFFT to align tip-rank-level sequences.
}
\details{
\code{"localpair"} selects the \bold{L-INS-i} algorithm, probably
  most accurate; recommended for <200 sequences; iterative refinement method
  incorporating local pairwise alignment information.

  \code{"globalpair"} selects the \bold{G-INS-i} algorithm suitable for
  sequences of similar lengths; recommended for <200 sequences; iterative
  refinement method incorporating global pairwise alignment information.

  \code{"genafpair"} selects the \bold{E-INS-i} algorithm suitable for
  sequences containing large unalignable regions; recommended for <200
  sequences.

  \code{"retree 1"} selects the \bold{FFT-NS-1} algorithm, the simplest
  progressive option in MAFFT; recommended for >200 sequences.

  \code{"retree 2"} selects the \bold{FFT-NS-2} algorithm that uses a second
  iteration of alignment based on a guide tree computed from an FFT-NS-1
  aligment; this is the default in MAFFT; recommended for >200 sequences.
}
\references{
Katoh, K. and H. Toh. 2008. Recent developments in the MAFFT
  multiple sequence alignment program. \emph{Briefings in Bioinformatics}
  \bold{9}: 286-298.

  Katoh, K., K.-i. Kuma, H. Toh, and T. Miyata. 2005. Mafft version 5:
  improvement in accuracy of multiple sequence alignment. \emph{Nucleic Acids
  Research} \bold{33}: 511--518.

  Katoh, K., K. Misawa, K.-i. Kuma, and T. Miyata. 2002. Mafft: a novel
  method for rapid multiple sequence alignment based on fast Fourier
  transform. \emph{Nucleid Acids Research} \bold{30}: 3059--3066.

  \url{http://mafft.cbrc.jp/alignment/software/index.html}
}
