% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepPASTA.R
\name{stepPASTA}
\alias{stepPASTA}
\title{STEP G: PASTA Alignment}
\usage{
stepPASTA(x, k = 200)
}
\arguments{
\item{x}{An object of class \code{\link{megapteraProj}}.}

\item{k}{An integer giving the size of a subset.}
}
\description{
Use PASTA to align tip-rank-level sequences.
}
\note{
This function is not fully implemented!
}
