% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepPOLENTA.R
\name{stepPOLENTA}
\alias{stepPOLENTA}
\title{STEP G: POLENTA Alignment}
\usage{
stepPOLENTA(x, k = 200)
}
\arguments{
\item{x}{An object of class \code{\link{megapteraProj}}.}

\item{k}{An integer giving the size of a subset.}
}
\description{
Use POLENTA to align tip-rank-level sequences.
}
\note{
This function is not fully implemented!
}
