% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip.infraspec.R
\name{strip.infraspec}
\alias{strip.infraspec}
\title{Truncation of Species Names}
\usage{
strip.infraspec(x)
}
\arguments{
\item{x}{A character string giving taxon names of species rank and below.}
}
\value{
A character string giving species names.
}
\description{
Delete infraspecific names and epithets form species names.
}
\details{
\code{strip.infraspec} tries to guess the separating character used;
  both \code{" "} and \code{"_"} are possible, but cannot be mixed as, e.g., 
  in \code{"Saxicola torquata_axillaris"}.
}
\examples{
strip.infraspec(c("Vipera aspis",
                  "Vipera aspis aspis",
                  "Vipera aspis atra"))
     
## separating characters cannot be mixed!
spec <- c("Vipera aspis",
          "Vipera_aspis_aspis",
          "Vipera_aspis atra")                               
strip.infraspec(spec) 
strip.infraspec(gsub(" ", "_", spec))
strip.infraspec(gsub("_", " ", spec))                   
}
