% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip.spec.R
\name{strip.spec}
\alias{strip.spec}
\title{Extract Genus Name from Linnean Binomial}
\usage{
strip.spec(x, mode = "strsplit")
}
\arguments{
\item{x}{A vector of mode \code{"character"} containing Linnean binomials.
The separating character can be space (\code{" "}) or underscore
(\code{"_"}).}

\item{mode}{A vector of mode \code{"character"} chosing between two
algorithms: \code{"strsplit"} is fast, but has failed on avery large vector
(> 170000 element); \code{"regex"} is not as fast, but has handled also
large vectors successfully.}
}
\value{
A vector of mode \code{"character"} containing genus names.
}
\description{
Strips the epitheton from a Linnean binomial and return the
  genus name.
}
\examples{
strip.spec("Megaptera_novaengeliae")

## separating characters cannot be mixed!
spec <- c("Megaptera_novaengeliae", "Megaptera novaeangliae")
strip.spec(spec) # does not work
strip.spec(gsub("_", " ", spec)) # this works
strip.spec(gsub(" ", "_", spec)) # this works
}
\seealso{
\code{\link{strip.infraspec}}
}
