% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxdump2phylo.R
\name{taxdump2phylo}
\alias{taxdump2phylo}
\title{Utilities for NCBI Taxdump}
\usage{
taxdump2phylo(tax, tip.rank)
}
\arguments{
\item{tax}{A data frame representing a taxonomy in parent-child format as 
returned by \code{\link{dbReadTaxonomy}}.}

\item{tip.rank}{A character string giving the name of a rank. This rank will be
treated as tip rank, i.e. all taxa of lower rank will be dicarded.}
}
\value{
An object of class \code{\link{phylo}}.
}
\description{
Convert a taxonomy table in parent-child format into an object 
  of class \code{\link{phylo}}.
}
\seealso{
\code{\link{dbReadTaxonomy}}, 
  \code{\link{taxdumpDaughters}},\code{\link{taxdumpLineage}}, 
  \code{\link{taxdumpSubset}}, \code{\link{taxdumpAddNode}}.
}
