% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxdump_isRoot.R
\name{taxdump_isRoot}
\alias{taxdump_isRoot}
\title{Utilities for NCBI Taxdump}
\usage{
taxdump_isRoot(x, taxon)
}
\arguments{
\item{x}{A data frame containing a parent-child table as returned by
\code{\link{dbReadTaxonomy}}.}

\item{taxon}{A character string giving the name of the node to be
tested.}
}
\description{
Check if node is root node.
}
\seealso{
\code{\link{ncbiTaxonomy}} for downloading the NCBI taxonomy and
  \code{\link{dbReadTaxonomy}} for reading the project taxonomy; other
  taxdump-related tools: \code{\link{taxdumpAddNode}},
  \code{\link{taxdumpChildren}}, \code{\link{taxdumpDropTip}},
  \code{\link{taxdumpHigherRank}}, \code{\link{taxdumpMRCA}},
  \code{\link{taxdumpSubset}}, \code{\link{taxdump2phylo}}
}
