% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxdump_isTerminal.R
\name{taxdump_isTerminal}
\alias{taxdump_isTerminal}
\title{Utilities for NCBI Taxdump}
\usage{
taxdump_isTerminal(tax, id)
}
\arguments{
\item{tax}{A data frame containing a parent-child table as returned by
\code{\link{dbReadTaxonomy}}.}

\item{id}{A vector of mode \code{"numeric"} giving the ID of the nodes to be
tested; can be missing an then all nodes are tested.}
}
\description{
Check if nodes are terminal, e.g. they are not parent nodes.
}
\seealso{
\code{\link{ncbiTaxonomy}} for downloading the NCBI taxonomy and
  \code{\link{dbReadTaxonomy}} for reading the project taxonomy; other
  taxdump-related tools: \code{\link{taxdumpAddNode}},
  \code{\link{taxdumpDaughters}}, \code{\link{taxdumpDropTip}},
  \code{\link{taxdumpHigherRank}}, \code{\link{taxdumpMRCA}},
  \code{\link{taxdumpSubset}}, \code{\link{taxdump2phylo}}
}
