% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxdumpAddNode.R
\name{taxdumpAddNode}
\alias{taxdumpAddNode}
\title{Utilities for NCBI Taxdump}
\usage{
taxdumpAddNode(x, tab, rank = "species", taxon, parent)
}
\arguments{
\item{x}{An object of class \code{\link{megapteraProj}} or a data frame in
parent-child format.}

\item{tab}{A data frame in parent-child format.}

\item{rank}{A character string giving the rank of the new taxon, default is
\code{"species"}.}

\item{taxon}{A character string giving the name of the new taxon.}

\item{parent}{A character string giving the name of the parent taxon.}
}
\value{
See Details section.
}
\description{
Insert a taxon as new node to the NCBI taxonomy.
}
\details{
\code{taxdumpAddNode} can be used in two ways: (1) If \code{x} is an
  object of class \code{"megapteraProj"}, the changes are made in the
  underlying postgreSQL database and \code{TRUE} or \code{FALSE} is returned
  indicating success or failure. (2) If \code{x} is a data frame in
  parent-child format, the changes are made to this table, which is then
  returned; in case of failure a warning is issued and the unchanged table is
  returned.
}
\seealso{
\code{\link{ncbiTaxonomy}} for downloading the NCBI taxonomy and
  \code{\link{dbReadTaxonomy}} for reading the project taxonomy; other
  taxdump-related tools: \code{\link{taxdumpChildren}},
  \code{\link{taxdumpDropTip}}, \code{\link{taxdumpHigherRank}},
  \code{\link{taxdumpMRCA}}, \code{\link{taxdumpSubset}},
  \code{\link{taxdump2phylo}} and \code{\link{taxdump_isTerminal}}.
}
