% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxdumpDropTip.R
\name{taxdumpDropTip}
\alias{taxdumpDropTip}
\title{Utilities for NCBI Taxdump}
\usage{
taxdumpDropTip(tax, id)
}
\arguments{
\item{tax}{Either an object of class \code{\link{megapteraProj}} or a data
frame containing a taxonomy table in parent-child format as returned by
\code{\link{dbReadTaxonomy}}.}

\item{id}{Numeric the ID of a terminal node (tip).}
}
\description{
Get subset of a taxonomy table in parent-child format.
}
\seealso{
\code{\link{dbReadTaxonomy}},
  \code{\link{taxdumpDaughters}},\code{\link{taxdumpLineage}},
  \code{\link{taxdumpAddNode}}, \code{\link{taxdump2phylo}}.
}
