% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxdumpHigherRank.R
\name{taxdumpHigherRank}
\alias{taxdumpHigherRank}
\title{Utilities for NCBI Taxdump}
\usage{
taxdumpHigherRank(x, taxon, rank)
}
\arguments{
\item{x}{Either an object of class \code{\link{megapteraProj}} or a data
frame containing a taxonomy table in parent-child format as returned by
\code{\link{dbReadTaxonomy}}.}

\item{taxon}{A vector of mode \code{"character"} giving the taxon name.}

\item{rank}{A vector of mode \code{"character"} giving the name of a higher (relative to
\code{taxon}) rank.}
}
\value{
A character string giving a taxon name.
}
\description{
Get the taxon name of a certain higher rank for a particular
  taxon, e.g. get the familiy of \emph{Megaptera}.
}
\seealso{
\code{\link{dbReadTaxonomy}},
  \code{\link{taxdumpDaughters}},\code{\link{taxdumpLineage}},
  \code{\link{taxdumpAddNode}}, \code{\link{taxdump2phylo}}.
}
