% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxdumpInfo.R
\name{taxdumpInfo}
\alias{taxdumpInfo}
\title{Utilities for NCBI Taxdump}
\usage{
taxdumpInfo(tax, taxon, lineage.depth)
}
\arguments{
\item{tax}{A data frame in parent-child format.}

\item{taxon}{A character string giving the name of the taxon to query.}

\item{lineage.depth}{Integer giving the maximum number of higher taxa when
displaying the lineage leading back to the root.}
}
\description{
Summarize information for a taxon name
}
\seealso{
\code{\link{taxdumpAddNode}}, \code{\link{taxdumpChildren}},
  \code{\link{taxdumpMRCA}}, \code{\link{taxdumpSubset}},
  \code{\link{taxdump2phylo}}
}
