% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxdumpManageSynonym.R
\name{taxdumpManageSynonym}
\alias{taxdumpManageSynonym}
\title{Utilities for NCBI Taxdump}
\usage{
taxdumpManageSynonym(
  tax,
  binomials,
  keep.syn = TRUE,
  add.syn = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{tax}{A data frame in parent-child format.}

\item{binomials}{A vector of mode \code{"character"} giving one or more taxon
names. The first name is considered to be the accepted name, the.}

\item{keep.syn}{Logical, should synonyms present in \code{tax} be kept?}

\item{add.syn}{Logical, should synonyms not present in \code{tax} be added?}

\item{quiet}{Logical, if \code{FALSE}, progress is printed on screen for
debugging.}
}
\description{
Manage Synonyms
}
\note{
\code{add.syn} is only partly implemented and not tested, because under
  the current use (adding synonyms to NCBI taxonomy) it does not make sense.
}
\seealso{
\code{\link{taxdumpAddNode}}, \code{\link{taxdumpChildren}},
  \code{\link{taxdumpMRCA}}, \code{\link{taxdumpSubset}},
  \code{\link{taxdump2phylo}}
}
