% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxdumpMergeNodes.R
\name{taxdumpMergeNodes}
\alias{taxdumpMergeNodes}
\title{Utilities for NCBI Taxdump}
\usage{
taxdumpMergeNodes(x, nodes, accept = nodes[1], new.child, new.child.rank)
}
\arguments{
\item{x}{An object of class \code{\link{megapteraProj}} or a data frame in
parent-child format.}

\item{nodes}{A vector of mode \code{"character"} giving the accepted names of
the nodes to be merged.}

\item{accept}{A character string giving the accpeted taxon name of the new node.}
}
\value{
See Details section.
}
\description{
Insert a taxon as new node to the NCBI taxonomy.
}
\details{
\code{taxdumpMergeNodes} can be used in two ways: (1) If \code{x} is an
  object of class \code{"megapteraProj"}, the changes are made in the
  underlying postgreSQL database and \code{TRUE} or \code{FALSE} is returned
  indicating success or failure. (\bold{not yet implemented!}) (2) If \code{x} is a data frame in
  parent-child format, the changes are made to this table, which is then
  returned.
}
\seealso{
\code{\link{ncbiTaxonomy}} for downloading the NCBI taxonomy and
  \code{\link{dbReadTaxonomy}} for reading the project taxonomy; other
  taxdump-related tools: \code{\link{taxdumpChildren}},
  \code{\link{taxdumpDropTip}}, \code{\link{taxdumpHigherRank}},
  \code{\link{taxdumpMRCA}}, \code{\link{taxdumpSubset}},
  \code{\link{taxdump2phylo}} and \code{\link{taxdump_isTerminal}}.
}
