% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxdumpMRCA.R
\name{taxdumpMRCA}
\alias{taxdumpMRCA}
\title{Utilities for NCBI Taxdump}
\usage{
taxdumpMRCA(x, species, tip.rank)
}
\arguments{
\item{x}{Either an object of class \code{\link{megapteraProj}} or a data
frame containing a taxonomy table in parent-child format as returned by
\code{\link{dbReadTaxonomy}}.}

\item{species}{A vector of mode \code{"character"} giving the species names.
Can be missing, in which case the root is returned.}

\item{tip.rank}{A character string giving the name a rank.}
}
\value{
A character string giving the name a the MRCA.
}
\description{
Get most-recent common ancestor (MRCA) of a group of species.
}
\seealso{
\code{\link{ncbiTaxonomy}} for downloading the NCBI taxonomy and
  \code{\link{dbReadTaxonomy}} for reading the project taxonomy; other
  taxdump-related tools: \code{\link{taxdumpAddNode}},
  \code{\link{taxdumpDaughters}}, \code{\link{taxdumpDropTip}},
  \code{\link{taxdumpHigherRank}}, \code{\link{taxdumpSubset}},
  \code{\link{taxdump2phylo}} and \code{\link{taxdump_isTerminal}}.
}
