% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxdumpRankOrder.R
\name{taxdumpRankOrder}
\alias{taxdumpRankOrder}
\title{Utilities for NCBI Taxdump}
\usage{
taxdumpRankOrder(x)
}
\arguments{
\item{x}{Either an object of class \code{\link{megapteraProj}} or a data
frame containing a taxonomy table in parent-child format as returned by
\code{\link{dbReadTaxonomy}}.}
}
\value{
A vector of mode \code{"character"} containing the ordered ranks.
}
\description{
Return the order of ranks in a taxonomy from highest to lowest.
}
\seealso{
\code{\link{ncbiTaxonomy}} for downloading the NCBI taxonomy and
  \code{\link{dbReadTaxonomy}} for reading the project taxonomy; other
  taxdump-related tools: \code{\link{taxdumpAddNode}},
  \code{\link{taxdumpChildren}}, \code{\link{taxdumpDropTip}},
  \code{\link{taxdumpMRCA}}, \code{\link{taxdumpSubset}},
  \code{\link{taxdump2phylo}} and \code{\link{taxdump_isTerminal}}.
}
