% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxdumpSynonym.R
\name{taxdumpSynonym}
\alias{taxdumpSynonym}
\title{Utilities for NCBI Taxdump}
\usage{
taxdumpSynonym(
  tax,
  binomials,
  user.acc,
  keep.acc = FALSE,
  keep.syn = TRUE,
  add.syn = TRUE,
  quiet = TRUE
)
}
\arguments{
\item{tax}{A data frame in parent-child format.}

\item{binomials}{A vector of mode \code{"character"} giving one or more taxon
names. The first name is considered to be the accepted name.}

\item{user.acc}{A vector of mode \code{"character"} giving the taxon names
that the user accepts.}

\item{keep.acc}{Logical, only applies when the accepted names in \code{tax}
and \code{binomial} are not identical: \code{TRUE} means the accepted name
from \code{tax} is kept, whereas the accepted name of \code{binomials} is
taken if \code{keep.acc = FALSE}.}

\item{keep.syn}{Logical, should synonyms present in \code{tax} be kept?}

\item{add.syn}{Logical, should synonyms not present in \code{tax} be added?}

\item{quiet}{Logical, if \code{FALSE}, progress is printed on screen for
debugging.}
}
\description{
Manage Synonyms
}
\note{
\code{add.syn} is only partly implemented and not tested, because under
  the current use (adding synonyms to NCBI taxonomy) it does not make sense.
}
\seealso{
\code{\link{taxdumpAddNode}}, \code{\link{taxdumpChildren}},
  \code{\link{taxdumpMRCA}}, \code{\link{taxdumpSubset}},
  \code{\link{taxdump2phylo}}
}
