% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitivityMerge.R
\name{transitivityMerge}
\alias{transitivityMerge}
\title{Transitivity Merge}
\usage{
transitivityMerge(obj, id = c(1, 2))
}
\description{
Merges two multiple sequence alignments using the transitivity 
  criterion.
}
