% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTips.R
\name{addTips}
\alias{addTips}
\title{Add Tips to a Phylogenetic Tree}
\usage{
addTips(phy, tax, tip.rank = "spec", tag = NULL, quiet = FALSE)
}
\arguments{
\item{phy}{An object of class \code{\link{phylo}}.}

\item{tax}{A character string giving the name of the 
species to be added.}

\item{tip.rank}{A character string giving the ranks of the 
tips in \code{phy}; must be present as a column name in 
\code{tax}.}

\item{tag}{\emph{To be added.}}

\item{quiet}{Logical, indicating if screen output should be 
suppressed.}
}
\value{
An object of class \code{\link{phylo}}.
}
\description{
Add tips (species) to a phylogenetic tree according 
to their taxonomic classification.
}
\examples{
data(cetacea)
tre <- addTips(cetacea$tree, cetacea$tax)
}
\seealso{
\code{\link{dbReadTaxonomy}} for reading a taxonomy table 
from the postgreSQL database.
}
