% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOLD2megaptera.R
\name{BOLD2megaptera}
\alias{BOLD2megaptera}
\title{Barcode of Life Database Interface}
\usage{
BOLD2megaptera(taxon, marker)
}
\arguments{
\item{taxon}{A character string giving a taxon name.}

\item{marker}{A character string giving a marker name.}
}
\description{
Get DNA sequences from BOLD SYSTEMS
}
