\name{checkIdentityCoverage}
\alias{checkIdentityCoverage}
\title{
Histograms of Identity/Coverage per Locus
}
\description{
Print histograms of the distribution of identity and coverage for each locus in the database to a PDF file.
}
\usage{
checkIdentityCoverage(megapteraProj)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{megapteraProj}{
an object of class \code{\link{megapteraProj}}.
  }
}
%\details{}
\value{none, as a side effect a PDF file is written to the working directory.}
%\references{}
\author{Christoph Heibl}
\seealso{
\code{\link{checkSpecLocus}},\code{\link{checkMissingSpec}}, \code{\link{checkSpecies}}.
}
%\examples{}
