% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbDNA.R, R/dbReadDNA.R, R/dbWriteDNA.R
\name{dbDNA}
\alias{dbDNA}
\alias{dbReadDNA}
\alias{dbWriteDNA}
\title{Read and Write DNA Sequences}
\usage{
dbReadDNA(x, tab.name, taxon, regex = TRUE, max.bp, min.identity,
  min.coverage, ignore.excluded = TRUE, subtree = FALSE,
  blocks = "ignore", reliability = 0)

dbWriteDNA(conn, tab.name, dna, enforce.binomial = TRUE, status)
}
\arguments{
\item{x}{An object of class \code{\link{megapteraProj}}.}

\item{tab.name}{A vector of mode \code{"character"} giving the name of the
database table. For \code{dbReadDNA} and when the locus is defined (see
\code{\link{setLocus}}) this can be omitted and the data will be read from
the species-level (or genus-level) database table.}

\item{taxon}{A vector of mode \code{"character"} used to choose a subset of
available taxa. This can be either one or more taxon names or a regular
expression.}

\item{regex}{Logical: if \code{TRUE}, the string given via \code{taxon} will
be interpreted as a regular expression (see \code{\link{regex}}).}

\item{max.bp}{An integer, only sequences equal or shorter than \code{max.bp}
will be returned.}

\item{min.identity}{A real number between 0 and 1, only sequences with a
fraction of at least \code{min.identity} nucleotides that are identical
with the reference sequence will be returned.}

\item{min.coverage}{A real number between 0 and 1, only sequences with a
fraction of at least \code{min.coverage} base pairs (compared to the
reference sequence) will be returned.}

\item{ignore.excluded}{Logical: if \code{FALSE}, \code{dbReadDNA} will also
return sequences that are marked in the \emph{status} field as 'excluded',
'too long', or 'too distant'.}

\item{subtree}{Logical.}

\item{blocks}{A vector of mode \code{"character"} indicating how to handle
alignment blocks: \code{"split"} causes blocks to be returned as elements
of a list. \code{"concatenate"} means, blocks will be concatendated and
returned as a single alignment. In order to get a (potentially unaligned)
list of sequences, choose \code{"ignore"}.}

\item{reliability}{A vector of mode \code{"numeric"} between 0 and 1 giving the
minimum reliability score for each alignment column.}

\item{conn}{A connection object as produced by \code{\link[DBI]{dbConnect}}.}

\item{dna}{An object of class \code{\link[ape]{DNAbin}} to write to the
database.}

\item{enforce.binomial}{Logical}

\item{status}{A vector of mode \code{"character"} to be written to the
\emph{status} field of the PostgreSQL table.}
}
\description{
Read and write DNA Sequences from/to PostgreSQL Database.
}
