% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbReadTaxomony.R, R/dbTaxonomy.R,
%   R/dbUpdateTaxonomy.R
\name{dbReadTaxonomy}
\alias{dbReadTaxonomy}
\alias{dbTaxonomy}
\alias{dbUpdateTaxonomy}
\title{Read and Write Taxonomy Table}
\usage{
dbReadTaxonomy(megProj, tip.rank, subset, tag, root = "tol",
  syn = FALSE)

dbUpdateTaxonomy(megProj, taxonomy, logfile = "")
}
\arguments{
\item{megProj}{An object of class \code{\link{megapteraProj}}.}

\item{tip.rank}{A character string giving the rank of the tips (e.g.
\code{species}, \code{genus}, ...).}

\item{subset}{A subset of species names (Latin binomials) to which the
taxonomy should be limited. Can be a DNA alignment of class \code{DNAbin},
a phylogenetic tree of class \code{phylo},\code{species_sequence} (see
Details), or a character vector.}

\item{tag}{A character string used to tag the rows of \code{taxonomy} in the
postgreSQL table; this argument is mostly for internal use.}

\item{root}{A character string. By default (\code{"tol"}),  all higher ranks
from the most recent common ancestor (MRCA) of the clade of interest down
to the root are included in the taxonomic classification, whereas
\code{"mrca"} means that only the MRCA will be included. Note that this
option only makes sense in combination with \code{subset}.}

\item{syn}{Logical, indicating if synonyms should be returned in addition to
accepted names.}

\item{taxonomy}{A data frame containing a taxonomic classification of the
study species. It will \bold{replace any existing} taxonomy in the
database.}

\item{logfile}{A character string giving the names of a potential logfile; no
logfile is written if this argument is left empty.}
}
\value{
For \code{dbReadTable} an object of class \code{"data.frame"} holding
  the taxonomic classification of target species. \code{dbUpdateTaxonomy} is
  called only for its side effect creating and extending the \code{taxonomy}
  database table.
}
\description{
Create, extend and read the \code{taxonomy} table of a
  \code{megaptera} project.
}
\details{
If step \code{\link{stepF}} has been run, \code{dbReadTaxonomy} can
  be called with \code{subset = "species_sequence"}. In that case all species
  names with sequence data for the current locus will be used for subsetting.
}
\seealso{
\code{\link{ncbiTaxonomy}} for retrieval of taxonomies from the
  taxonomy database at NCBI; \code{\link{findRoot}} returns the lowest common
  rank for taxonomies. \code{\link{dbReadDNA}} and \code{\link{dbWriteDNA}}
  to read and write DNA sequences.
}
