\name{megaptera-internal}

\alias{conn}
\alias{tax}
\alias{loci}
\alias{pars}
\alias{addNbAcc}
\alias{alignGenus}
\alias{calcDistToRef}
\alias{check.TaxAgainstPhy}
\alias{compare.phylo.backbone}
\alias{coverage}
\alias{dbAutoFillTaxonomy}
\alias{dbconnect}
\alias{dbDeleteSpec}
\alias{dbReadDNAbin}
\alias{dbReadReference}
\alias{dbUpdateLocus}
\alias{dbWriteDNAbin}
\alias{dbWriteString}
\alias{downloadSequences}
\alias{EFetchHigherTaxon}
\alias{EFetchLocus}
\alias{esearch}
\alias{extractLocus}
\alias{filter.alignment}
\alias{fixTaxonomy} % deserves explicit doc
\alias{force.mp}
\alias{funX}
\alias{getGeneRange}
\alias{getLineage}
\alias{getMaxDist}
\alias{insert.tip.phylo}
\alias{is.alignable}
\alias{MAD}
\alias{mafft.merge}
\alias{markersGenbank}
\alias{maxDist}
\alias{maxDistMPI}
\alias{meanPairwiseIdentity}
\alias{megapteraHeatmap}
\alias{myDist}
\alias{ncbiLineage}
\alias{NCBI.wrap}
\alias{nucleotideOverlap}
\alias{read.benchmark.db}
\alias{revCompTest}
\alias{select.seqs}
\alias{slog}
\alias{specCons}
\alias{speciesConsensus}
\alias{sortRanks}
\alias{splitGiTaxon}
\alias{sql.conform}
\alias{sqlTaxonomyHeader}
\alias{sql.wrap}
\alias{stepAA}
\alias{summary.sproj}
\alias{supermatrix}
\alias{tab}
\alias{tax2tree}
\alias{term}
\alias{unalignable}
\alias{whereToInsert}
\alias{write.dna.spectable}


\title{Internal MEGAPTERA Functions}
\description{Internal \pkg{megaptera} functions.}
\note{
These are internal functions in \code{megaptera} and are not intended to be called by the user.
}
\keyword{internal}
