\name{megaptera-package}
\alias{megaptera-package}
\docType{package}
\title{
MEGAPhylogenie TEchniques in R
}
\description{
The goal of developing \bold{megaptera} is the automation of sequence-retrieval from internet databases (currently Nucleotide at GenBank) to produce (possibly large) aligned nucleotide sequence datasets of one or more genetic markers \bold{for comparative analyses}.
}
\details{
\tabular{ll}{
Package: \tab megaptera\cr
Type: \tab Package\cr
Version: \tab 1.0-0\cr
 Date: \tab 2014-10-30\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Christoph Heibl
Maintainer: Christoph Heibl <christoph.heibl@gmx.net>
}
\references{
Nuleotide at NCBI: \url{http://www.ncbi.nlm.nih.gov/nuccore}
}
\keyword{ package }
\seealso{
\code{\link[ape:ape-package]{ape}}
}
%\examples{}
