% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/megapteraProj-class.R
\docType{class}
\name{megapteraProj-class}
\alias{megapteraProj-class}
\title{Class "megapteraProj"}
\description{
S4 class for holding all input data of the megaptera pipeline.
}
\section{Slots}{

\describe{
\item{\code{db}}{An object of class \code{\linkS4class{dbPars}}.}

\item{\code{taxon}}{An object of classes \code{\linkS4class{taxon}} or
\code{\linkS4class{taxonGuidetree}}.}

\item{\code{locus}}{An object of classes \code{\linkS4class{locus}} or
\code{\linkS4class{locusRef}}.}

\item{\code{align.exe}}{A vector of mode \code{"character"}, giving name of the
alignment program; currently only \bold{MAFFT} is allowed.}

\item{\code{merge.exe}}{A vector of mode \code{"character"}, giving name of the
alignment merging program; currently only \bold{OPAL} is allowed.}

\item{\code{mask.exe}}{A vector of mode \code{"character"}, giving name of the
alignment masking program; currently only \bold{Gblocks} is allowed.}

\item{\code{params}}{An object of class \code{\linkS4class{megapteraPars}}.}

\item{\code{update}}{Logical: if \code{TRUE}, the pipeline's steps are executed as
if called for the first time, i.e., possibly overriding data and setting
that have been previously achieved.}
}}

\seealso{
\code{\link{dbPars}}, \code{\link{taxon}}, \code{\link{locus}}, and
  \code{\link{megapteraPars}} for defining of database parameters, taxa,
  loci, and the pipeline's parameters, respectively.
}
