% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/megapteraProj.R
\name{megapteraProj}
\alias{megapteraProj}
\title{Bundle Input Data for the Pipeline}
\usage{
megapteraProj(db, taxon, locus = locus(), align.exe = "undefined",
  merge.exe = "undefined", mask.exe = "undefined",
  params = megapteraPars(), update = FALSE)
}
\arguments{
\item{db}{An object of class \code{\linkS4class{dbPars}}.}

\item{taxon}{An object of classes \code{\linkS4class{taxon}} or
\code{\linkS4class{taxonGuidetree}}.}

\item{locus}{An object of classes \code{\linkS4class{locus}} or
\code{\linkS4class{locusRef}}.}

\item{align.exe}{A vector of mode \code{"character"}, giving name of the
alignment program; currently only \bold{MAFFT} is allowed.}

\item{merge.exe}{A vector of mode \code{"character"}, giving name of the
alignment merging program; currently only \bold{OPAL} is allowed.}

\item{mask.exe}{A vector of mode \code{"character"}, giving name of the
alignment masking program; currently only \bold{Gblocks} is allowed.}

\item{params}{An object of class \code{\linkS4class{megapteraPars}}.}

\item{update}{Logical: if \code{TRUE}, the pipeline's steps are executed as
if called for the first time, i.e., possibly overriding data and setting
that have been previously achieved.}
}
\value{
An object of class \code{\linkS4class{megapteraProj}}.
}
\description{
Create an object of class \code{\linkS4class{megapteraProj}} to
  hold all the input data necessary for running of the pipeline.
}
\references{
MAFFT: \url{http://mafft.cbrc.jp/alignment/software/}

  OPAL: \url{http://opal.cs.arizona.edu/}

  Gblocks: \url{http://molevol.cmima.csic.es/castresana/Gblocks.html}
}
\seealso{
\code{\link{dbPars}}, \code{\link{taxon}},
  \code{\link{taxonGuidetree}}, \code{\link{locus}}, \code{\link{locusRef}},
  and \code{\link{megapteraPars}} for defining of database parameters, taxa,
  loci, and the pipeline's parameters, respectively.
}
