% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbiTaxonomy.R
\name{ncbiTaxonomy}
\alias{ncbiTaxonomy}
\title{NCBI Taxonomy Database}
\usage{
ncbiTaxonomy(x, quiet = FALSE)
}
\arguments{
\item{x}{Database connection parameters, either as object of class
\code{\link{dbPars}} or \code{\link{megapteraProj}}.}

\item{quiet}{Logical, indicating if diagnostic messages should be printed on
screen.}
}
\value{
\code{ncbiTaxonomy} is called for its side effect (see Details).
}
\description{
Retrieve taxonomic classification from the taxonomy database at
  the National Center for Biotechnology Information (NCBI).
}
\details{
The NCBI taxonomy database will be downloaded via FTP in "taxdump"
  format, unpacked, and translated into data frames. In a second step, the
  data frames will be stored in a postgreSQL database called
  \code{"ncbitaxonomy"}. Any existing data will be overwritten in the
  process.
}
\references{
NCBI Taxonomy Database website:
  \url{http://www.ncbi.nlm.nih.gov/taxonomy}

Federhen, Scott. 2012. The NCBI taxonomy database. \emph{Nucleic
  Acids Research} \bold{40}: DI36-DI43.
}
\seealso{
\code{\link{stepA}} for extracting the taxonomic information
  relevant for a given megaptera project; \code{\link{dbUpdateTaxonomy}} and
  \code{\link{dbReadTaxonomy}} for storing and retrieving taxonomic
  information in a megaptera project database;
  \code{\link{dbSummaryTaxonomy}} for a short numerical summary of the
  project taxonomy.
}
