% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbiTaxonomy.R
\name{ncbiTaxonomy}
\alias{ncbiTaxonomy}
\title{NCBI Taxonomy Database}
\usage{
ncbiTaxonomy(x)
}
\arguments{
\item{x}{Database connection parameters, either as objdect of class 
\code{\link{dbPars}} or \code{\link{megapteraProj}}.}
}
\value{
\code{ncbiTaxonomy} is called for its side effect (see Details).
}
\description{
Retrieve taxonomic classifications from the taxonomy database at
  the National Center for Biotechnology Information (NCBI).
}
\details{
The NCBI taxonomy database will be downloaded via FTP in "taxdump" 
  format, unpacked, and translated into a data frame object. In a second 
  step, the data frame table will be stored in a postgreSQL database called 
  \code{"ncbitaxonomy"}.
}
\references{
Taxonomy Database website:
  \url{http://www.ncbi.nlm.nih.gov/taxonomy}

Federhen, Scott. 2012. The NCBI taxonomy database. \emph{Nucleic
  Acids Research} \bold{40}: DI36-DI43.
}
\seealso{
\code{\link{stepA}} for extracting the taxonomic information 
  relevant for a given \bold{megaptera} project.
  \code{\link{dbUpdateTaxonomy}} and \code{\link{dbReadTaxonomy}} for storing
  and retrieving taxonomic information in a \bold{megaptera} project
  database.
}
