% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPhyloWithClassification.R
\name{plotPhyloWithClassification}
\alias{plotPhyloWithClassification}
\title{Plot Phylogentic Tree with Clade Highlighting}
\usage{
plotPhyloWithClassification(x, phy, og, surrogate.tips, min.support,
  collapse.unsupported = TRUE, ignore.tag, edges, higher.rank,
  pdf.height = "auto", pdf.width = 20, title, file)
}
\arguments{
\item{x}{An object of class \code{\link{megapteraProj}}.}

\item{phy}{An object of class \code{\link{phylo}}.}

\item{og}{A vector of mode \code{"character"}.}

\item{collapse.unsupported}{Logical}

\item{ignore.tag}{\emph{Currently unused.}}
}
\description{
Plots a phylogenetic tree together with color highlighting of
  the underlying taxonomic classification.
}
