% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.phylo.rank.R
\name{prune.phylo.rank}
\alias{prune.phylo.rank}
\title{Prune Phylogenies to Higher Ranks}
\usage{
prune.phylo.rank(phy, tax, rank = "genus")
}
\arguments{
\item{phy}{An object of class \code{\link[ape]{phylo}}.}

\item{tax}{A data frame containing taxonomic information for the tip labels
in \code{phy}.}

\item{rank}{A character string giving the name of a column (= taxonomic rank)
in \code{tax} to which \code{phy} will be pruned.}
}
\value{
An object of class \code{\link[ape]{phylo}}.
}
\description{
Prune tips of a certain taxonomic rank from an object of class
  \code{\link[ape]{phylo}} and obtain a new \code{phylo} object whose tips
  are of a higher rank.
}
\seealso{
\code{\link{addTips}} and \code{\link{addSingleTip}} to add terminal nodes to a phylogeny.
}
