% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speclist2taxdump.R
\name{speclist2taxdump}
\alias{speclist2taxdump}
\title{Utilities for NCBI Taxdump}
\usage{
speclist2taxdump(species, root = "root")
}
\arguments{
\item{species}{A vector of mode \code{"character"} giving a set of species or
genus names.}

\item{root}{A character string giving the taxon name of the root node.}
}
\description{
Convert a list of species or genera into a taxonomy table in
  parent-child format.
}
\seealso{
\code{\link{dbReadTaxonomy}},
  \code{\link{taxdumpChildren}},\code{\link{taxdumpLineage}},
  \code{\link{taxdumpAddNode}}, \code{\link{taxdump2phylo}}.
}
