% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepA.R
\name{stepA}
\alias{stepA}
\title{Step A: Creating a Project Taxonomy}
\usage{
stepA(x)
}
\arguments{
\item{x}{An object of class \code{\link{megapteraProj}}.}
}
\value{
None. \code{stepA} is called for its side effects: (1) a taxonomic 
  classifiaction is stored in the pgSQL database; (2) a log file is written 
  to the \code{log} directory of the project directory.
}
\description{
Creates a project taxonomy from the NCBI taxonomy (see 
  \code{\link{ncbiTaxonomy}}).
}
\note{
Any subsequent call to \code{stepA} will overwrite the existing
  \code{taxonomy} table. As a consequence, \code{stepB} and following steps
  will have to be rerun also.
}
\seealso{
\code{\link{megapteraProj}} for bundling the project's input 
  information; \code{\link{stepB}} for the next step in the pipeline.
}
