% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepF.R
\name{stepF}
\alias{stepF}
\title{Step F: Select Sequences and Assemble FASTA file}
\usage{
stepF(x, update)
}
\arguments{
\item{x}{An object of class \code{\link{megapteraProj}}.}

\item{update}{Logical, if \code{FALSE}, step F is redone from scratch, if
\code{TRUE}, updating is done only if the data or parameters have changed. If
left empty, \code{update} is taken from \code{\link{dbPars}}.}
}
\description{
In \code{stepF} FASTA files will be assembled selecting all
  sequences that passed the quality evalution.
}
\seealso{
\code{\link{megapteraProj}}; \code{\link{stepE}} for the preceeding
  and \code{\link{stepG}} for the subsequent step.
}
