\name{stepF}
\alias{stepF}
\title{Step F: Select Sequences and Assemble FASTA file}
\description{
In \code{stepF} FASTA files will be assembled selecting all sequences that passed the quality evalution.
}
\usage{
stepF(x)
}
\arguments{
  \item{x}{an object of class \code{\link{megapteraProj}}.}
}
%\details{}
%\value{}
%\references{}
\author{
Christoph Heibl
}
\note{
The function is under development and the results should always be checked.
}
\seealso{
\code{\link{megapteraProj}}; \code{\link{stepE}} for the preceeding and \code{\link{stepG}} for the subsequent step.
}
%\examples{}
