% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepG.R
\name{stepG}
\alias{stepG}
\title{Step G: Profile Alignment}
\usage{
stepG(x)
}
\arguments{
\item{x}{An object of class \code{\link{megapteraProj}}.}
}
\description{
Performs a profile alignment (with MAFFT) of the sequences
  selected by \code{\link{stepD}} along a taxonomy build with \code{stepA}.
}
\note{
The function is under development and the results should always be
  checked.
}
\seealso{
\code{\link{megapteraProj}}; \code{\link{stepF}} for the preceeding
  and \code{\link{stepH}} for the subsequent step.
}
