% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxdumpAddNode.R
\name{taxdumpAddNode}
\alias{taxdumpAddNode}
\title{Utilities for NCBI Taxdump}
\usage{
taxdumpAddNode(x, taxon, rank, parent)
}
\arguments{
\item{x}{An object of class \code{\link{megapteraProj}}.}

\item{taxon}{A character string giving the name of the new taxon.}

\item{rank}{A character string giving the rank of the new taxon.}

\item{parent}{A character string giving the name of the parent taxon.}
}
\value{
None, \code{taxdumpAddNote} is called for its side effect of
  inserting data in the underlying PostgreSQL database.
}
\description{
Insert a taxon as new node to the NCBI taxonomy.
}
\seealso{
\code{\link{ncbiTaxonomy}} for downloading the NCBI taxonomy and
  \code{\link{dbReadTaxonomy}} for reading the project taxonomy; other
  taxdump-related tools: \code{\link{taxdumpChildren}},
  \code{\link{taxdumpDropTip}}, \code{\link{taxdumpHigherRank}}, \code{\link{taxdumpMRCA}},
  \code{\link{taxdumpSubset}}, \code{\link{taxdump2phylo}} and
  \code{\link{taxdump_isTerminal}}.
}
