% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxdumpChangeParent.R
\name{taxdumpChangeParent}
\alias{taxdumpChangeParent}
\title{Utilities for NCBI Taxdump}
\usage{
taxdumpChangeParent(x, taxon, new.parent)
}
\arguments{
\item{x}{An object of class \code{\link{megapteraProj}} or a \code{data.frame}
a parent-child-format.}

\item{taxon}{A character string giving the name of the taxon.}

\item{new.parent}{A character string giving the name of the new parent taxon.}
}
\value{
A \code{data.frame} in parent-child-format, if \code{class(x) ==
  "data.frame"}, or \code{TRUE}, if \code{class(x) == "megapteraProj"}. In
  the latter case \code{taxdumpChangeParent} is called for its side effect on
  the taxonomy table of the database.
}
\description{
Change a node's parent in the NCBI taxonomy.
}
\seealso{
\code{\link{ncbiTaxonomy}} for downloading the NCBI taxonomy and
  \code{\link{dbReadTaxonomy}} for reading the project taxonomy; other
  taxdump-related tools: \code{\link{taxdumpChildren}},
  \code{\link{taxdumpDropTip}}, \code{\link{taxdumpHigherRank}},
  \code{\link{taxdumpMRCA}}, \code{\link{taxdumpSubset}},
  \code{\link{taxdump2phylo}} and \code{\link{taxdump_isTerminal}}.
}
