% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxdumpDaughters.R
\name{taxdumpDaughters}
\alias{taxdumpDaughters}
\title{Utilities for NCBI Taxdump}
\usage{
taxdumpDaughters(tax, taxon, immediate = FALSE, tip.rank = "species",
  indet = indet.strings())
}
\arguments{
\item{tax}{Either an object of class \code{\link{megapteraProj}} or a data
frame as returned by \code{\link{dbReadTaxonomy}}.}

\item{taxon}{A character string giving the name of the taxon.}

\item{immediate}{Logical, if \code{TRUE}, only the immediate children of
\code{taxon} will be returned. The default (\code{FALSE}) return all
children.}

\item{tip.rank}{A character string giving the name a rank. This rank will be
treated as tip rank, i.e. all taxa of lower rank will be dicarded.}

\item{indet}{A vector of character strings containing regular expressions
(see Examples).}
}
\value{
A data frame.
}
\description{
Get all children of certain rank for a given taxon.
}
\examples{
# The set of default regular expressions used to identify nonvalid species binomials
indet.strings()
}
\seealso{
\code{\link{taxdumpLineage}}, \code{\link{taxdumpMRCA}},
  \code{\link{taxdump2phylo}}.
}
