% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxdumpDropTip.R
\name{taxdumpDropTip}
\alias{taxdumpDropTip}
\title{Utilities for NCBI Taxdump}
\usage{
taxdumpDropTip(tax, id)
}
\arguments{
\item{tax}{Either an object of class \code{\link{megapteraProj}} or a data
frame containing a taxonomy table in parent-child format as returned by
\code{\link{dbReadTaxonomy}}.}

\item{id}{a vector of mode \code{"numeric"}, giving one or more IDs of a
terminal node (tip).}
}
\description{
Get subset of a taxonomy table in parent-child format.
}
\seealso{
\code{\link{ncbiTaxonomy}} for downloading the NCBI taxonomy and
  \code{\link{dbReadTaxonomy}} for reading the project taxonomy; other
  taxdump-related tools: \code{\link{taxdumpAddNode}},
  \code{\link{taxdumpChildren}}, \code{\link{taxdumpHigherRank}},
  \code{\link{taxdumpMRCA}}, \code{\link{taxdumpSubset}},
  \code{\link{taxdump2phylo}} and \code{\link{taxdump_isTerminal}}.
}
