% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxdumpHigherRank.R
\name{taxdumpHigherRank}
\alias{taxdumpHigherRank}
\title{Utilities for NCBI Taxdump}
\usage{
taxdumpHigherRank(x, taxon, rank, it.max = 30)
}
\arguments{
\item{x}{Either an object of class \code{\link{megapteraProj}} or a data
frame containing a taxonomy table in parent-child format as returned by
\code{\link{dbReadTaxonomy}}.}

\item{taxon}{A vector of mode \code{"character"} giving the taxon name.}

\item{rank}{A vector of mode \code{"character"} giving the name of a higher
(relative to \code{taxon}) rank.}

\item{it.max}{An integer giving maximum number of iteration in case
\code{rank} is not present in the lineage of \code{taxon}}
}
\value{
A character string giving a taxon name.
}
\description{
Get the taxon name of a certain higher rank for a particular
  taxon, e.g. get the familiy of \emph{Megaptera}.
}
\seealso{
\code{\link{ncbiTaxonomy}} for downloading the NCBI taxonomy and
  \code{\link{dbReadTaxonomy}} for reading the project taxonomy; other
  taxdump-related tools: \code{\link{taxdumpAddNode}},
  \code{\link{taxdumpChildren}}, \code{\link{taxdumpDropTip}},
  \code{\link{taxdumpMRCA}}, \code{\link{taxdumpSubset}},
  \code{\link{taxdump2phylo}} and \code{\link{taxdump_isTerminal}}.
}
