% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxdumpLineage.R
\name{taxdumpLineage}
\alias{taxdumpLineage}
\title{Utilities for NCBI Taxdump}
\usage{
taxdumpLineage(tax, taxon, highest.rank)
}
\arguments{
\item{tax}{Either an object of class \code{\link{megapteraProj}} or a data
frame as returned by \code{\link{dbReadTaxonomy}}.}

\item{taxon}{A character string giving the name of the taxon.}

\item{highest.rank}{A character string giving the highest rank returned; if
missing (default) the highest rank returned is the root.}
}
\value{
A data frame.
}
\description{
Get all higher ranks including a given taxon.
}
\details{
If \code{taxon} is coded as a synonym, it will be replaced by the
  corresponding accepted name and a warning will be issued.
}
\seealso{
\code{\link{taxdumpAddNode}}, \code{\link{taxdumpChildren}},
  \code{\link{taxdumpMRCA}}, \code{\link{taxdumpSubset}},
  \code{\link{taxdump2phylo}}
}
