% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxdumpLineage.R
\name{taxdumpLineage}
\alias{taxdumpLineage}
\title{Utilities for NCBI Taxdump}
\usage{
taxdumpLineage(tax, taxon)
}
\arguments{
\item{tax}{Either an object of class \code{\link{megapteraProj}} or a data
frame as returned by \code{\link{dbReadTaxonomy}}.}

\item{taxon}{A character string giving the name of the taxon.}
}
\description{
Get all higher ranks including a given taxon.
}
\seealso{
\code{\link{taxdumpAddNode}}, \code{\link{taxdumpChildren}},
  \code{\link{taxdumpMRCA}}, \code{\link{taxdumpSubset}},
  \code{\link{taxdump2phylo}}
}
