% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxdumpSanity.R
\name{taxdumpSanity}
\alias{taxdumpSanity}
\title{Sanity Check for Parent-Child Taxonomic Tables}
\usage{
taxdumpSanity(tax)
}
\arguments{
\item{tax}{A data frame in parent-child format.}
}
\value{
Logical
}
\description{
Does several sanity checks for taxonomic tables in parent-child format.
}
\details{
\code{taxdumpSanity} checks, if
\enumerate{
\item synonyms have the same parent as their corresponding accepted name
\item accepted species names the corresponding genus name assigned
}
}
