% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxdumpSubset.R
\name{taxdumpSubset}
\alias{taxdumpSubset}
\title{Utilities for NCBI Taxdump}
\usage{
taxdumpSubset(tax, mrca, species, root = "tol", syn = FALSE)
}
\arguments{
\item{tax}{Either an object of class \code{\link{megapteraProj}} or a data
frame containing a taxonomy table in parent-child format as returned by
\code{\link{dbReadTaxonomy}}.}

\item{mrca}{A character string giving the most recent common ancestor (MRCA)
of the subset.}

\item{species}{A vector of mode \code{"character"} giving a subset of species
names.}

\item{root}{A character string choosing between \code{"mrca"} and
\code{"tol"} (tree of life).}

\item{syn}{Logical, indicating if synonyms should be returned in addition to
accepted names; only has an effect if \code{tax} is of class
\code{\link{megapteraProj}}.}
}
\description{
Get subset of a taxonomy table in parent-child format.
}
\seealso{
\code{\link{dbReadTaxonomy}},
  \code{\link{taxdumpChildren}},\code{\link{taxdumpLineage}},
  \code{\link{taxdumpAddNode}}, \code{\link{taxdump2phylo}}.
}
